pyproject.toml
[tool.poetry]
name = "salarycalc"
version = "2.0"
description = "Python scripts to update Salary Calc base values."
classifiers= [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "License :: OSI Approved :: MIT License",
]
authors = ["Niteo <info@niteo.co>"]
homepage = "https://niteo.co/salary-calculator"
keywords = [ "salaries"]
license = "MIT"
packages = [
    { include = "salarycalc", from = "scripts" },
]

[tool.poetry.dependencies]
python = "^3.11"

"ruamel.yaml" = "*"
requests = "*"
tqdm = "*"

# This version must match the playwright-browsers version in $PLAYWRIGHT_BROWSERS_PATH
playwright = "==1.47.0"

[tool.poetry.dev-dependencies]
black = "*"
codespell = "*"
flake8 = "*"
flake8-assertive = "*"
flake8-blind-except = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-deprecated = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-mock = "*"
flake8-mutable = "*"
flake8-plone-hasattr = "*"
flake8-self = "*"
flake8-super-call = "*"
flake8-todo = "*"
flake8-tuple = "*"
mypy = "*"
pdbpp = "*"
pre-commit-hooks = "*"
pydocstyle = "<4.0.0"
yamllint = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
